足すときに一時的にMOD超えるため,配列はlonglongで定義

+++ AtCoder/Educational_DP_Contest/P_Independent_Set.cpp
@@ -71,7 +71,7 @@ struct UnionFind
 
 vector<vector<int>> a;
 
-int dp[100010][2];
+ll dp[100010][2];
 bool flag[100010];
 
 void f(int node){
